requirements.txt
# Above Exterior Cleaning — Business Intelligence Hub
# Install: pip install -r requirements.txt

# Core
python-dotenv>=1.0.0
requests>=2.31.0

# Supabase
supabase>=2.4.0
psycopg2-binary>=2.9.9

# QuickBooks Online
intuit-oauth>=1.2.6
python-quickbooks>=0.9.4

# Google APIs (Business Profile + Gmail)
google-api-python-client>=2.120.0
google-auth>=2.28.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0

# Meta / Facebook Ads
facebook-business>=19.0.0

# Slack
slack-bolt>=1.18.0

# AI / Anthropic
anthropic>=0.23.0

# Scheduling
APScheduler>=3.10.4
SQLAlchemy>=2.0.0

# Utilities
python-dateutil>=2.9.0
